rename PickSpoiler.py to pick_spoiler.py
(and update all calls to it)
