+++ hints/gui/main_page.py
@@ -1,7 +1,7 @@
 
 # Home to the base functions based on the main page.
 
-from hints.gui.PickSpoiler import spoiler_pop_up
+from hints.gui.pick_spoiler import spoiler_pop_up
 from hints.gui.Utils import create_notebook_tab, create_scrollable
 from tkinter import Button, Frame, Tk
 from tkinter.ttk import Notebook
